TEIID-4546 TEIID-4547 fixing various issues with domain mode and get query plans

+++ admin/src/main/java/org/teiid/adminapi/Admin.java
@@ -507,7 +507,7 @@ public interface Admin {
      * @param executionId
      * @return
      */
-    String getQueryPlan(String sessionId, int executionId) throws AdminException;
+    String getQueryPlan(String sessionId, long executionId) throws AdminException;
 
     /**
      * Restart the Server
